kernel_config: Fix net configset splitting
Before this commit, it was not possible to boot CLIP OS without the
net/basic configset. Although it certainly does not make a lot of sense
to do so, it is cleaner now.

Note that we already had CONFIG_NET in basic at it is needed by other
non-network-related stuff.

The tricky part was that CONFIG_CRYPTO_AUTHENC (which was selected by
some IPsec Kconfig options) seems to actually be required by dm-crypt,
as otherwise we get a "crypt: Error allocating crypto tfm" error when
cryptsetup tries to open the encrypted volume. This is an upstream
limitation [1].

[1] https://lore.kernel.org/lkml/20190920154434.GA923@gandi.net/t/#u

Change-Id: I5df15c3f98d3cdc6ed011f6baa12f3d47fa02219
Signed-off-by: Thibaut Sautereau <thibaut.sautereau@ssi.gouv.fr>